[Chore] rename `menuService.ts` to `menu.ts`
