+++ src/platform/menu/common/menu.register.ts
@@ -1,7 +1,7 @@
 import { createRegister, RegistrantType } from "src/platform/registrant/common/registrant";
 import { MenuTypes, IMenuItemRegistration } from "src/platform/menu/browser/menuRegistrant";
 import { IS_MAC } from "src/base/common/platform";
-import { CommandID } from "src/platform/menu/common/menuService";
+import { CommandID } from "src/platform/menu/common/menu";
 
 // Application Menu (For macOS)
 export const menuTitleApplicationRegister = createRegister(
