+++ src/platform/menu/electron/mainMenuService.ts
@@ -4,7 +4,7 @@ import { ILogService } from "src/base/common/logger";
 import { IpcChannel } from "src/platform/ipc/common/channel";
 import { SafeIpcMain } from "src/platform/ipc/electron/safeIpcMain";
 import { IMenuItemRegistrationResolved, MenuTypes } from "src/platform/menu/browser/menuRegistrant";
-import { IMenuService } from "src/platform/menu/common/menuService";
+import { IMenuService } from "src/platform/menu/common/menu";
 import { IMainWindowService } from "src/platform/window/electron/mainWindowService";
 
 const mainMenuTypes: { type: MenuTypes; label: string }[] = [
